Migrate move related structs to move_def.hpp
