+++ src/game.hpp
@@ -8,115 +8,12 @@
 #include <SDL_image.h>
 #include <SDL_timer.h>
 
+#include "move_def.hpp"
+
 using namespace std;
 
 
-enum Piece {
-    WHITE_KING = 'K',
-    WHITE_QUEEN = 'Q',
-    WHITE_ROOK = 'R',
-    WHITE_BISHOP = 'B',
-    WHITE_KNIGHT = 'N',
-    WHITE_PAWN = 'P',
-    BLACK_KING = 'k',
-    BLACK_QUEEN = 'q',
-    BLACK_ROOK = 'r',
-    BLACK_BISHOP = 'b',
-    BLACK_KNIGHT = 'n',
-    BLACK_PAWN = 'p'
-};
-
-
-struct MoveLog {
-    ushort index;
-    ushort x1;
-    ushort y1;
-    ushort x2;
-    ushort y2;
-    char piece1;
-    char piece2;
-};
-
-
-class Move {
-    public:
-        vector<vector<int>> possible;
-        vector<vector<int>> controlled;
-
-        void p_add(int x, int y){
-            possible.push_back({x, y});
-        }
-
-        void p_add(int square[2]){
-            possible.push_back({square[0], square[1]});
-        }
-
-        void c_add(int x, int y){
-            controlled.push_back({x, y});
-        }
-
-        void c_add(int square[2]){
-            controlled.push_back({square[0], square[1]});
-        }
-};
-
-
-class MoveInfo{
-    public:
-        Move all_moves[8][8];
-
-        MoveInfo(){
-            for (int x = 0; x < 8; x++){
-                for (int y = 0; y < 8; y++){
-                    all_moves[x][y] = Move();
-                }
-            }
-        }
-
-        void clear(){
-            for (int x = 0; x < 8; x++){
-                for (int y = 0; y < 8; y++){
-                    all_moves[x][y] = Move();
-                }
-            }
-        }
-
-        Move take(int x, int y){
-            return all_moves[x][y];
-        }
-
-        Move take(int piece[2]){
-            return all_moves[piece[0]][piece[1]];
-        }
-
-        void put(int x, int y, Move move){
-            all_moves[x][y] = move;
-        }
-
-        void put(int piece[2], Move move){
-            all_moves[piece[0]][piece[1]] = move;
-        }
-
-        int p_size(int x, int y){
-            return all_moves[x][y].possible.size();
-        }
-
-        int p_size(int piece[2]){
-            return all_moves[piece[0]][piece[1]].possible.size();
-        }
-
-        int c_size(int x, int y){
-            return all_moves[x][y].controlled.size();
-        }
-
-        int c_size(int piece[2]){
-            return all_moves[piece[0]][piece[1]].controlled.size();
-        }
-};
-
-
-class Tester;
-
+class Tester; // unit testing class - friend of Game
 
 class Game{
     private:
